Fix LaunchPad synchronization
Inside the container we need to mark the git folder manually as trusted
to allow `git` commands to run.
Additionally the SSH-setup action uses the $HOME directory while we run
as root in the container and hence SSH uses `/root/.ssh` and ignores `$SSH_PATH`.
So copy the prepared configuration to that path.