Fix launchpad upload
Use a pip installed breezy and apply patches to make it work.
The first is a bug in the code (missing import) the 2nd an
incompatibility with Python3 where `sys.stdin` is a string buffer while
a byte buffer is expected. `sys.stdin.buffer` can be used for that as
done in later breezy versions.